+++ lib/offsite_payments/integrations/nochex.rb
@@ -98,7 +98,7 @@ module OffsitePayments #:nodoc:
           raise ArgumentError, "amount must be a Money object or an integer" if money.is_a?(String)
           raise ActionViewHelperError, "amount must be greater than $0.00" if cents.to_i <= 0
 
-          add_field mappings[:amount], sprintf("%.2f", cents.to_f/100)
+          add_field mappings[:amount], sprintf("%.2f", cents.to_d/100)
         end
 
         # Optional Parameters
@@ -180,7 +180,7 @@ module OffsitePayments #:nodoc:
 
         # the money amount we received in X.2 decimal.
         def gross
-          sprintf("%.2f", params['amount'].to_f)
+          sprintf("%.2f", params['amount'].to_d)
         end
 
         # Was this a test transaction?
